requirements.txt
fuzzywuzzy==0.18.0
openpyxl==3.1.5
pandas==2.3.3
psycopg2==2.9.2
python-dotenv==1.2.1
Requests==2.32.5
Unidecode==1.4.0
